Updated names in lab1
