+++ code/korniakov-kirill-lab1/src/test/java/ru/unn/agile/complexnumber/model/AndreevVladislavTest.java
@@ -2,7 +2,7 @@ package ru.unn.agile.complexnumber.model;
 
 import org.junit.Test;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 
 
 public class AndreevVladislavTest {
