+++ code/korniakov-kirill-lab1/src/test/java/ru/unn/agile/complexnumber/model/ChekodaevOlegTest.java
@@ -1,7 +1,8 @@
 package ru.unn.agile.complexnumber.model;
 
 import org.junit.Test;
-import static org.junit.Assert.*;
+
+import static org.junit.Assert.assertEquals;
 
 public class ChekodaevOlegTest {
     @Test
