+++ code/korniakov-kirill-lab1/src/test/java/ru/unn/agile/complexnumber/model/ComplexNumberTest.java
@@ -1,6 +1,7 @@
 package ru.unn.agile.complexnumber.model;
 
 import org.junit.Test;
+
 import static org.junit.Assert.*;
 
 public class ComplexNumberTest {
