+++ code/korniakov-kirill-lab1/src/test/java/ru/unn/agile/complexnumber/model/RumiantsevaMariaTest.java
@@ -1,6 +1,7 @@
 package ru.unn.agile.complexnumber.model;
 
 import org.junit.Test;
+
 import static org.junit.Assert.assertEquals;
 
 public class RumiantsevaMariaTest {
