+++ code/korniakov-kirill-lab1/src/test/java/ru/unn/agile/complexnumber/model/ShiryaevaEkaterinaTest.java
@@ -1,7 +1,9 @@
 package ru.unn.agile.complexnumber.model;
 
 import org.junit.Test;
-import static org.junit.Assert.*;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
 
 public class ShiryaevaEkaterinaTest {
     @Test
